+++ src/ts/constant.ts
@@ -1,0 +1,16 @@
+/**
+ * Peek.js
+ * 
+ * A lightweight JavaScript library that attaches a viewer to a specific node type, allowing you to view the CSS properties, attributes, and size/position.
+ * 
+ * @file        constant.ts
+ * @version     v1.1.0
+ * @author      Bunoon
+ * @license     MIT License
+ * @copyright   Bunoon 2024
+ */
+
+
+export namespace Constant {
+    export const PEEK_JS_IGNORE_STATE_ATTRIBUTE = "data-peek-js-ignore-state";
+}