+++ src/ts/dom.ts
@@ -11,7 +11,8 @@
  */
 
 
-import { Char } from "./enum";
+import { Constant } from "./constant";
+import { Char, IgnoreState } from "./enum";
 import { Is } from "./is";
 import { type Position } from "./type";
 
@@ -22,6 +23,8 @@ export namespace DomElement {
         const isText: boolean = nodeType === "text";
 
         let result: any = isText ? document.createTextNode( Char.empty ) : document.createElement( nodeType );
+
+        result.setAttribute( Constant.PEEK_JS_IGNORE_STATE_ATTRIBUTE, IgnoreState.ignore );
 
         if ( Is.defined( className ) ) {
             result.className = className;
@@ -35,6 +38,7 @@ export namespace DomElement {
     export function createWithHTML( container: HTMLElement, type: string, className: string, html: string ) : HTMLElement {
         const element: HTMLElement = create( container, type, className );
         element.innerHTML = html;
+        element.setAttribute( Constant.PEEK_JS_IGNORE_STATE_ATTRIBUTE, IgnoreState.ignore );
 
         return element;
     }